+++ packages/signers/pjs-signer/src/index.ts
@@ -1,3 +1,3 @@
 export * from "./injected-extensions"
-export { getTxCreatorFromPjs, type PjsTxCreator } from "./from-pjs-account"
+export { getTxCreatorFromPjs } from "./from-pjs-account"
 export type { SignerPayloadJSON, SignPayload, SignRaw } from "./types"